+++ src/assets/svgs/closed-issue.ts
