+++ src/assets/svgs/index.ts
@@ -1,5 +1,5 @@
-export * from "./Commit";
-export * from "./OpenedIssue";
-export * from "./ClosedIssue";
-export * from "./OpenedPull";
-export * from "./MergedPull";
+export * from "./commit";
+export * from "./opened-issue";
+export * from "./closed-issue";
+export * from "./opened-pull";
+export * from "./merged-pull";
