+++ src/assets/svgs/merged-pull.ts
