+++ src/assets/svgs/opened-issue.ts
