+++ src/assets/svgs/opened-pull.ts
